emacs conf: remove SPC & DEL from raghu/scroll-map
The keymap is meant for scrolling one line at a time while keeping
point on the original text-line.  Using SPC / DEL to scroll by
screenfuls is not in agreement with the intention of the map.  Using
SPC / DEL to scroll by one line is against established convention of
scrolling with SPC/DEL, which is by screenfuls (page-wise).  So, just
remove SPC and DEL from the map.  Use the usual key-bindings `C-v' and
`M-v' instead to scroll a page at a time.