emacs conf: ease up on `let'
In the body of `let', it is OK to use `setq' on the symbols bound in
that `let' form's varlist.

I was overdoing `let', and all that excessive scoping was making it
hard to implement some parts without bugs, in addition to making
things a little hard-to-follow.  Specifically, I was unable to
implement in a bug-free manner the parts where I account for
save-excursion's behavior at the beginning of the line.  The excessive
scoping was causing me to lose the information I need to decide how
many lines to skip ahead.  With the changes in this commit, I am doing
only the required scoping, and using `setq' safely, so I have all the
information I need to decide correctly how many lines to skip ahead.