emacs conf: unbind C-x C-z & C-z; use remap vector
Using the remap vector makes it clearer that I mean "when I type any
of the key-bindings for list-buffers, run ibuffer instead."

C-x z followed by z repeats the previous command, and that's
sufficiently easy that I do not need C-z.  I never used C-x C-z, and
having it bound to the default (suspend-frame) is an inconvenience
because I fat-finger more than I like to admit.