vim: clear VertSplit highlighting before linking
Using a single `highlight!' seems to give VertSplit its original
highlighting, and the link to Special, but only the link to Special is
used.  So, do it more cleanly(?) by first clearing VertSplit's
highlighting, and then linking to Special.  I have no idea if this extra
step is necessary; just seemed cleaner to do it this way.