vimrc: do not put `if...endif' all in one line
Split it up into multiple lines.

Vimscript parsing when if...endif is all in one line is different
compared to what I had imagined.

----

	if exists(":packadd") | packadd! matchit | endif

When the above is run on any version of Vim that does not support the
packadd command, Vim waits for the command to be completed, as if there
is no endif at the end even though it is right there.

	if exists(":packadd")
		packadd! matchit
	endif

When split up into multiple lines like above, the behavior matches my
expectation: the command packadd does not exist, the check fails, and so
nothing happens.

----

	if exists(":packadd") | foo | endif

If the above is run on a version of Vim which supports the packadd
command, the following error message is displayed:

	E492: Not an editor command: foo | endif

I expected only `foo' to show up there, and not `foo | endif' (`foo' is
a contrived example of an invalid Vim command).

	if exists(":packadd")
		foo
	endif

The above, however, complains as expected about only `foo' (right after
evaluation of line 2), rather than about `foo' and `endif'.

----

So, always split if...endif across multiple lines, and reduce confusion
for myself.