vimrc: discard xsel stderr
If xsel fails, I do not want the buffer to get modified with xsel's
stderr.  So, discard the stderr output.  If the command fails, Vim will
show the message 'shell returned <status>' in the command line.  Coupled
with the fact that the expected action did not occur in the buffer, that
message is sufficient notification for me that the command failed.