Vim: update color scheme 'raghuvrao'
Make diffAdded/diffRemoved subtler.

Color the background when sufficent colors are available.  Otherwise,
color just the foreground with the basic colors (colors likely to be
available on most terminal types), and leave the background alone.

Modify some highlight groups' colors for better contrast

Use has('gui_running') instead of has('gui') to decide when to modify
colors pertinent to the GUI.

Pop-up menu color updates.

Add StatusLine, StatusLineNC and VertSplit: even on terminals with just
8 colors, there should be good contrast now.  The default highlighting
did Gray ("normal" white) on dark gray (bright black) because of the
bold attribute on the status line foreground.  The overall effect was
poor contrast in the Linux console, which is something that always
bothered even though I use a terminal emulator in X.  This commit should
put and end to this botheration.  The default for VertSplit is
reverse-video, which is OK for the most part, but I added it too, so
things do not look too mismatched.

Modify visual selection when the background is dark: change it so that
it is not the same color as the status line for the window containing
the cursor.