raghuvrao_dflb0.vim: various updates
Add StatusLine, StatusLineNC and VertSplit: even on terminals with just
8 colors, there should be good contrast now.  The default highlighting
did Gray ("normal" white) on dark gray (bright black) because of the
bold attribute on the status line foreground.  The overall effect was
poor contrast in the Linux console, which is something that always
bothered even though I use a terminal emulator in X.  This commit should
put and end to this botheration.  The default for VertSplit is
reverse-video, which is OK for the most part, but I added it too, so
things do not look too mismatched.

Modify visual selection when the background is dark: change it so that
it is no the same color as the status line for the window containing the
cursor.