vimrc: make it easier to spot current window
When no highlighting is available, Vim identifies the window containing
the cursor (the active window) by filling the empty areas of that
window's status line with "^", and the empty areas of the other windows'
status lines with "=".  If highlighting is available, Vim does not fill
the empty areas of the status lines with these characters.  Arrange to
make Vim display these characters regardless of the availability of
highlighting: it will become that much easier to tell the active window
from the others when I have several splits.