raghuvrao_dflb0.vim: ditch dark-bg colors
When I originally created this colorscheme, I created it with the
intention of using it only when the background is light and the
foreground is dark.  I even put 'dflb' (dark foreground, light
background) in its name.  At some point, I forgot about the original
intent, and added stuff for dark-background-light-foreground too.

So, remove the dark-background-related lines.