Vim: update color scheme 'raghuvrao'
Do not split lines.

In some cases, when Vim is invoked as the command "ex", there are error
messages about lines beginning with the reverse solidus.  I _think_ it
happens when ex is a symbolic link to vi, and vi is Vim with
Vi-compatible behavior (:set compatible).  So, do not split the
highlight-specification lines across multiple lines.