raghuvrao_dflb0.vim: :hi clear terminal-statusline
Before linking terminal statusline highlights to regular statusline
highlights, clear the terminal statusline highlights.  Otherwise, the
terminal statusline hightlights will have the defaults AND the link (but
still use only the link), which seems a little weird to me.  I am not
sure if clearing the highlights is really necessary, so this commit is
mostly only so I can stop obsessing.