raghuvrao_dflb0.vim: change order of attributes
Go from the following order:

	term,
	cterm, ctermfg, ctermbg,
	gui, guifg, guibg

to the following order:

	ctermfg, ctermbg, cterm
	guifg, guibg, gui
	term

Sometimes, the bold attribute does not get set when using the former
order.  It seems to happen when ctermfg is one of the Dark variants
(see :h cterm-colors), and cterm=bold is set before ctermfg.

:h cterm-colors has the following (Vim 7.4.1938):

If you want to set the bold attribute in a different way, put a
"cterm=" argument AFTER the "ctermfg=" or "ctermbg=" argument.	Or use
a number instead of a color name.