raghuvrao_dflb0.vim: netrw highlights
netrwClassify: it makes sense to me to keep this one highlighted, so it
is clearer that the classification character is not part of the
file-name.  So, define highlights for it.

netrwDir, netrwSymLink: remove highlighting for these two.  The
netrwClassify highlight is sufficient for me to tell if a file is
a directory or a symbolic link.