Vim: stricter check for tmux-related settings
To prevent the escape-sequences from being assigned in GUI-Vim (the
assignments are unnecessary there), the 'gui_running' check is needed:
'term' is not yet set to 'builtin_gui' when GUI-Vim runs the commands
vimrc.  Instead, 'term' will have the same value as the shell variable
TERM; it gets set to 'builtin_gui' some time later.