bashrc, bash_profile: add shebang line for clarity
The files bashrc and bash_profile are both meant to be sourced by shells,
and not to be used as commands.  So, a shebang line is not necessary in
either file.  Regardless, add shebang lines to these files because doing
so allows some text editors to load file-type specific text-editing
functions correctly.