bashrc, profile: add shebang line for clarity
The files bashrc and profile are both meant to be sourced by shells, and
not to be used as commands.  So, a shebang line is not necessary in
either file.  Regardless, add shebang lines to these files, only to
emphasize further that the file bashrc is meant for bash, and the file
profile is expected to be sh-compatible.

A positive side-effect:

For the file profile, Emacs and Vim set the major-mode and filetype
correctly now, respectively.  Previously (without the shebang line),
Emacs would use fundamental-mode and Vim would set the filetype to conf
for profile.

When I access the file profile through the path ~/.profile (symbolic
link) in Vim, Vim sets the correct filetype, whereas accessing it
through the real path/filename ('profile') would set the filetype to
conf.  The reason is fairly evident from $VIMRUNTIME/filetype.vim.
I assume that filetype.vim does not regard a file named 'profile' as
a shell script because the word 'profile' is far too common to make
assumptions about its filetype.

I have configured Emacs to resolve symbolic links always, so even when
accessing the file through the path ~/.profile, Emacs would not load the
correct major-mode without the shebang line.

By adding the shebang lines, the above minor inconvenience is resolved
in both editors.