raghuvrao_dflb0.vim: change status line colors
Bold+White/DarkCyan looks good only when bold (actual boldface, and not
just via brighter colors) is available, AND bright variants are used for
bold.  In the text console, sometimes, bright variants are not
available.  For example, when I am using a console font that has more
than some specific number of glyphs (128?), and if framebuffer is not
available, I do not get bold-via-bright-colors.  And because it is the
console, there is no actual boldface.  In this case, Bold+White/DarkCyan
will be rendered as LightGray/DarkCyan, which is unreadable.

Using Bold+Cyan/DarkBlue for the current window's status line and
LightGray/DarkBlue for non-current-window's status line seems to work
better.