raghuvrao_dflb0.vim: Make VertSplit FG blue too
...but a lighter shade of blue.  In this way, on the Linux console
(which has a dark background and light foreground), the foreground text
of the vertical split (the vertical bar) and the text right next to the
vertical split are set apart better.  It looks quite good on
light-background-dark-foreground screens too, in my opinion at least.

For color text-terminals, make the reasonable assumption that Blue and
DarkBlue are different, and that the former is a lighter shade compared
to the latter.