tmux: use non-zero escape-time
A small delay will allow tmux to recognize meta- and function-key
sequences of the terminal more reliably.  I have not really felt the
difference (probably because most of the time, tmux is running locally,
and not on a remote machine), but set escape-time to a small delay
anyway because it is safer.