tmux.conf: improvement to message-command-style
	set -g message-command-style fg=brightyellow,bg=brightblack

In the GNU/Linux console, this setting makes the letter under the cursor
appear in dark gray (brightblack) foreground on yellow (brown)
background, which is rather unreadable.

	set -g message-command-style fg=black,bg=brightyellow,reverse

In the GNU/Linux console, this setting makes the letter under the cursor
appear in black foreground on yellow (brown) background, which is much
more readable for me.

In terminal emulators, this setting and the one above it both result in
the same appearance.  So, this commit is only for improving the
appearance of message-command-style in the GNU/Linux console, which does
not seem to be able display the "bright" variants as background colors.