Do not silence the terminal bell
Let programs ring the terminal bell as they please.

When/if I get sick of the bell, I will silence it in one place: in the
configuration of the terminal (or the terminal emulator).

Silencing the bell per-program is annoying.  I have never felt the need
to enable the bell for some programs and keep it disabled for others; it
has always been enable-everywhere or disable-everywhere for me.