inputrc/readline/bash: add a few mappings
Occasionally, I use vi-style mappings in bash.

Add the following mapping to the vi-command keymap:

`o' to re-read the readline init file (~/.inputrc).  Helps when I want
to re-read ~/.inputrc into the current shell process after making
changes to the file.  I could not find a default mapping for this
function in the vi-command keymap.

Add the following mapping to the vi-insert keymap:

CTRL-L to clear the screen.  The default mapping is in the vi-command
keymap, and so, ESC CTRL-L (or, equivalently, CTRL-META-L) is required
for clearing the screen.  CTRL-L was unused (well, self-insert) in
vi-insert, so I am going to use CTRL-L for clear-screen.

Add the following mapping to both vi-command and vi-insert keymaps:

CTRL-_ (CTRL-/) to place the final argument of the previous command into
the current line.  Frequently, I find myself running a command with some
argument, and then running another command with the same argument, but
with slight modifications.  This mapping will help to recall that
argument quickly.  The advantage over `!!' is that I can edit the
argument before running the command.