bash_profile: function declarations: put `{' on own line
`declare -f' puts the opening brace for a function on its own line.
I realize I am trying to keep this file sh-compatible, and `declare' is
a bash builtin.  Still, I am taking this recommendation from `declare'
and applying it to this file because I like this recommendation.

A neat side-effect is that Vim properly indents shell functions whose
names contain one or more hyphens.