Vim: when filetype is set, ensure textwidth != 0
Some ftplugins define textwidth, and others do not.  For those filetypes
whose ftplugins do not define textwidth, set textwidth to 79.  I am
making certain that textwidth is never zero for any filetype mostly to
affect the behaviour of the gq operator.  When textwidth is 0, gq
formats text to 79 columns or the number of columns in the Vim window,
whichever is smaller.  Instead, I want gq to format to within 79 columns
regardless of the Vim window size, unless the filetype's ftplugin says
otherwise.  If the buffer does not have a filetype, do nothing about
textwidth.  See :help gq.