Vim: set/reset list more aggressively
I want 'list' to be set in all buffers except buffers where the filetype
is either 'help' (Vim help) or 'man' (ftplugin/man.vim).

Unfortunately for me, 'list' is a window-local option, and not
buffer-local, which causes 'list' to be changed when I do not want it to
be changed, when switching buffers.