Vim: allow ftplugins to change concealing settings
Set conceallevel and concealcursor only at the global level.  Doing so
allows ftplugins to modify these options per filetype.  Do not enforce
the same values everywhere.

Doing so reduces confusion in a few filetypes.  For example, in Vim help
buffers.  Look at the following excerpt from `:help fo-table', with
conceallevel=0 and empty concealcursor.

	...
2	When formatting text, use the indent of the second line of a paragraph
	for the rest of the paragraph, instead of the indent of the first
	line.  This supports paragraphs in which the first line has a
	different indent than the rest.  Note that 'autoindent' must be set
	too.  Example: >
			first line of a paragraph
		second line of the same paragraph
		third line.
<	This also works inside comments, ignoring the comment leader.
v	Vi-compatible auto-wrapping in insert mode: Only break a line at a
	...

In the above, it looks like the `<' character is an option and the
sentence following it is describing it, whereas that sentence is part of
the description for `2'.  The `<' character (and the `>' character) in
the above excerpt is actually a concealed character.