Vim: (Vim 8) move cursor to position of previous exit
I upgraded to Vim 8!

I also uninstalled the Vim packages provided by Slackware.  In doing so,
I had to give up /usr/share/vim/vimrc, which had the autocommand that,
upon reading a file into a buffer, makes the cursor jump to the position
in which I was when I last edited that buffer.  So, add that
functionality into my ~/.vimrc, with my modifications to it.