vimrc: do not include byte offset in statusline
I am doing it incorrectly.

The way I had it, Vim kept appending the byte offset to the statusline
each time I moved a Go buffer back into a window.  Probably because
'statusline' is window-local, and not buffer-local.

Remove the augroup until I find the time+patience to figure it out.