vimrc: fix+simplify how I set 'textwidth'
Doing it in an autocommand is a bit risky: when 'binary' is set,
'textwidth' is set to 0, which the autocommand will change to 79 when
I move the buffer out of the window and back in.  This behaviour is
undesirable.  Setting just the global value for 'textwidth' is
sufficient.  Some ftplugins modify textwidth only if it is 0, and all of
these plugins seem to be setting it to 78, which coincidentally, is the
value to which I am setting it right now.