Vim: update color scheme 'raghuvrao'
Blue for MoreMsg and Question.

On a light-background terminal where only 8 colors are available (the
background is assumed to be white, which is actually gray), Blue (either
bold/bright or dark) offers better contrast than DarkGreen.  I want
Question and MoreMsg to be legible even in such situations where colors
are very limited in availability.