Vim: update color scheme 'raghuvrao'
Comment: use Blue instead of DarkBlue in cterm; fallback to DarkBlue if
bold will be forced to get Blue from DarkBlue.  Blue offers more
contrast against the default foreground color I use than DarkBlue
offers, and Blue looks legible on the default background color I use.

Ignore: added.  I want to be able to disable concealing, and be able to
tell between Ignore-highlighted text and regular text.

NonText: bold when in GVim to make it easier to differentiate between
'#' at the end of a column in a diffRemoved-highlighted line, and the
'precedes'/'extends' listchars.  In cterm, use just DarkMagenta because
the DarkRed and DarkMagenta colors in terminal emulators are fairly
well-contrasted.

Todo: do not use boldface in cterm.  In many terminal emulators and
terminals, boldface implies the bright variant of the color.  If the
default background is gray, Magenta text becomes hard to read.
DarkMagenta, on the other hand, remains legible.

gitcommitOverflow: highlight gitcommitSummary instead.

helpSpecial, helpHyperTextEntry, helpHyperTextJump: use foreground
colors that offer better contrast on a gray background, should I ever
feel the need to switch to a gray background.