Vim: update color scheme 'raghuvrao'
Add Normal, but only for GVim.

Mostly for the sake of completeness.  I will not modify the default
BG/FG colors in the terminal / terminal-emulator, however.  Sometimes,
the exact colors I want are unavailable; sometimes, there is the
background-color-erase problem; sometimes, there is something else, all
of which annoy me immensely.