Vim: prevent undesired automatic scrolling
When I type zt (or zb) in the current buffer (i.e. move the view such
that the current line is at the top (or bottom) of the window), then
switch to another buffer in the same window, and then switch back to the
first buffer in the same window, Vim centers the current line in the
window instead of leaving the line at the top (or bottom).  Add
functions and autocommands to work around this problem.