GVim: LightGray for Ignore instead of LightYellow
LightYellow foreground on White background makes the foreground
practically indistinguishable from the background.  Even though it is
Ignore-highlighted, I want to be able to see the text.  So, use
LightGray for the foreground which seems to have just the right amount
of contrast with White so as to be only subtly visible.