vimrc: remove an unnecessary condition
For buffers created by the :help command, 'buftype' will be set
automatically to 'help'.  The condition &buftype !=# '' (which is
checked before the file type) will therefore be true making the entire
condition true.

When I open a Vim help file to edit it, the file type can be 'help', but
the 'buftype' will be '' (i.e. a normal buffer), and in this case,
I want the cursor to get moved on initial load.

So, remove the &filetype ==# 'help' condition.