Vim: update color scheme 'raghuvrao'
Refuse to load when <8 colors.

I need at least 8 colors (for the UI elements).

If I can get 16, it is better because then I can use the bright / light
variants for the background too, and not just the foreground, which is
important for a colorscheme meant for use on light-background screens.