zshrc: rebind C-U to backward-kill-line
By default, zsh binds C-U to kill-whole-line, which is something I have never
needed to use.  I just use C-C to cancel a line if I have made a mess of it.
backward-kill-line (kill from point to beginning of line) is a far more useful
thing for me, so bind it to C-U (which also makes the functionality of C-U in
zsh similar to bash).