Delete pythonrc.py
I added this file on 2013-09-19, in the very first commit in this repository,
with the intention of using it some day.  Since then (in all likelihood, since
_before_ then), I have not actually used this file.  More than 6 years have
passed now.  Plus, in Python3, whatever I am doing in this file is already
done by default.  And, I do not write so much Python these days as Go.  So,
remove this file.