Vim: remove SaveView and RestoreView
I got the ideas for these functions from:

https://vim.fandom.com/wiki/Avoid_scrolling_when_switch_buffers

There is still some undesired scrolling, and no matter what I do, I cannot
seem to get rid of it.  Remove these functions, and adapt to this Vim
idiosyncrasy because fighting it is not worth my time and energy.  Most of the
time, I use multiple Vim-windows in one screen anyway, so I do not have to
deal with the problem I was trying to solve with these functions.