Vim: show more information in the statusline
Show percentage through file of the displayed window.  Useful to know if
the end of the file is visible in the window without having to move the
cursor to check.

Show the file format.  Useful to know the kind of line-endings the file
being edited has, in case line-endings for one platform cause problems
on another platform.