Emacs: improve+split raghu/duplicate-and-comment
I learned that `line-beginning-position' and `line-end-position' take
line counts as arguments, which lets me avoid a few `save-excursion'
calls.

Split up `raghu/duplicate-and-comment' into
`raghu/duplicate-and-comment-region' that acts only on the region, and
`raghu/duplicate-and-comment-lines' that acts only on a line count.
Replace the original function with
`raghu/duplicate-and-comment-lines-or-region' that calls one of the two
functions as appropriate.  Feels more maintainable this way.