Emacs: set fewer env variables
Do not bother with LESS.  I do not set it in my environment outside
Emacs, so it is not necessary to unset it within Emacs' environment.

Do not bother with GIT_PAGER, FCEDIT, GIT_EDITOR, and SVN_EDITOR.  The
programs that use these variables fall back to EDITOR or VISUAL when
these variables are not defined.  Outside Emacs, I do not have these
variables, so no need to unset them in Emacs.