Emacs: set fewer env variables, but more carefully
Do not bother with LESS.  I do not set it in my environment outside
Emacs, so it is not necessary to unset it within Emacs' environment.

Do not bother with GIT_PAGER, FCEDIT, GIT_EDITOR, and SVN_EDITOR.  The
programs that use these variables fall back to EDITOR or VISUAL when
these variables are not defined.  Outside Emacs, I do not have these
variables anymore, so no need to unset them in Emacs.

When setting EDITOR and VISUAL, use absolute paths, and specific ones
too, without regard to exec-path or PATH.  Feels more secure this way.