Emacs: remove count-lines usage; somewhat fix save-excursion hack
`save-excursion' appears to do the equivalent of creating a marker that
does not advance on insert when the marker is at the beginning of line.
In some cases, I want the marker to advance on insert, regardless of
where the marker originally was.  In these cases, create and manage the
marker myself rather than using `save-excursion'.

I was using `count-lines' in a few functions solely to return the
number, or to workaround the `save-excursion' behaviour described above.
Now that I am not using `save-excursion' where I must work around its
behaviour, and I have never really used the return value of some
functions (I use them interactively for side-effects, and ignore the
return value), remove `count-lines' from them.