ZLE, Readline: add mappings to insert previous last arguments
In the 'emacs' keymap, M-. inserts the previous command's last argument.
I find it convenient, and there is no default mapping for it in the
vi-insert keymaps in both ZLE and Readline.  So add a mapping.  From
what I can tell, C-_ (which seems to be the same as C-/) is not mapped
to any ZLE or Readline function in the vi-insert keymap.