Readline: make CTRL-L clear the screen in vi-insert mode
In Readline, CTRL-L is bound to clear-screen in the emacs and vi
keymaps, and not in the vi-insert keymap.  I find it slightly
inconvenient to switch to the vi-command (vi) mode while in vi-insert
mode just to clear the screen.  So, bind CTRL-L to clear-screen in the
vi-insert keymap too.