bash_profile, zprofile: do not bother with PAGER
When PAGER is undefined, most programs use 'less'.  And, I was setting
PAGER to 'less'.  In the case of some programs, if PAGER is not set,
they use a specific invocation of 'less' which I lose by setting PAGER.
So, let programs decide how they want to invoke a pagination program.