Vim: enable 'hlsearch' instead of the toggle
Enable 'hlsearch', and map :nohlsearch instead of :set invhlsearch so
I do not have toggle 'hlsearch'.  I thought 'hlsearch' was too
distracting, so I had kept it disabled to toggle only when needed.
I find myself enabling it more than disabling it, and when I disable it,
it is a minor inconvenience to re-enable it.  :nohlsearch solves this
problem.