urxvt/Xresources: blink the cursor
Most of the time, I like a blinking cursor to a steady one.  The
exception so far is xterm because xterm insists on blinking the cursor
even as I am typing / moving the cursor around, which makes for an
unpleasant experience for me.  urxvt's approach to blinking the cursor
is more sensible, so enable it.