zsh/bash/readline: indicate vi mode in prompt
In Bash, it is mostly straightforward.  The change to PS1 in bashrc is
to remain consistent with what I have in Zsh.

In Zsh, getting the vi mode indicator to work correctly in the prompt
has a few problems, all of which stem from the reset-prompt widget,
which unfortunately is inevitable to get the prompt to update when the
keymap changes.

When ZLE is using vi mappings, AND PS1 is a multiline string, AND the
zle-keymap-select hook runs the reset-prompt widget, typing M-a, M-A,
M-i, M-I, M-o, or M-O (which, in my terminal emulator, are ESC followed
by the corresponding letter) will cause ZLE to erase a line above.
I find this bug/feature exceedingly annoying.

To solve this problem, I can either use a single line prompt or use
something other than ESC to go from viins/main to vicmd, thus making ESC
<letter> nothing special.  The latter approach is not feasible for me,
so using a single-line prompt seemed like the solution.  Unfortunately,
the reset-prompt widget also has the effect of erasing any partial lines
of output (lines that do not end in a new line), and it happens
regardless of how many lines the prompt spans; I will never see the
output of 'echo -n hello'.

So, I use this solution: move the bulk of the "prompt" into precmd(),
and make PS1 (and PS2) extremely simple and one line.

The disadvantage of this precmd() approach is that when I clear the
screen, the output of precmd() is also erased, and I have to type RETURN
to get ZLE to print the prompt, and in doing so, run precmd().  I think
I am willing to live with this disadvantage for now; having a working vi
mode indicator without the problems described above is probably worth
enduring this minor inconvenience.