Vim: make JumpToLastExitPosition() more efficient
Most of the time, the condition in the if-clause would evaluate to false
for me, which means all of the conditions had to be checked, because the
conditions are OR'd.  By switching them to ANDs, many of the checks will
be skipped most of the time, making this function finish a little faster
most of the time.