Bash: make PATH cleanup/rebuild more efficient
At least, I think it is more efficient because it seems more efficient.
I have not benchmarked it.  The shell's startup time is still
instantaneous for me.

Instead of repeated string comparisons and modifications, use an array
and an associative array to remove duplicates from PATH.  Then, use the
array to reconstruct PATH, without having to worry about a leading or
trailing ':'.

This approach eliminates the need for the path_append function, so
remove it.

A leading or trailing colon in PATH means an empty entry at,
respectively, the beginning or end of PATH.  An empty entry in PATH
means the current directory is included in PATH at that position.
Including the current directory in PATH is a security risk.