Vim: add mappings to move to start/end of words and stay there
b, e, B, and E move to consecutive words upon consecutive use. I would
like a motion similar to ^ but for words: a mapping that moves the
cursor to the start/end of words, and consecutive uses of this mapping
keep the cursor at the start/end of the same word.