Zsh: make M-f and C-u more useful
Bind M-f to emacs-forward-word which moves the cursor to the end of
words; the default forward-word moves the cursor to the beginning of
words. So, now, with M-f and M-b, I have a way to go to the beginning
and end of words, and not just the beginning of words.

By default, C-u is bound to kill-whole-line, which is not very useful.
I can just use C-c. Instead, bind C-u to backward-kill-line (kill from
the beginning of the line to the cursor position), which acts as the
counterpart for C-k (kill from cursor position to end of line), and it
is far more useful than kill-whole-line.