Emacs: face customizations
If the changes being made to a face is not making that face nil, then make the
change if at least 256 colors are available to Emacs (and when the background is
light); otherwise, let Emacs use the defaults for that face. In this way, I can
worry less about which colors are available.